% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_clusters_within_shift.R
\name{average_clusters_within_shift}
\alias{average_clusters_within_shift}
\title{Average Clusters within Shift (Embedded in find_niches)}
\usage{
average_clusters_within_shift(start_tile, clusters, tile_shift, tile_height)
}
\arguments{
\item{tile_height}{}
}
\description{
Average Clusters within Shift (Embedded in find_niches)
}
