% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{back_data}
\alias{back_data}
\title{Background environmental data for examples}
\format{
A data frame with 12,396 rows and 4 variables:
\describe{
  \item{x}{Longitude in decimal degrees (WGS84).}
  \item{y}{Latitude in decimal degrees (WGS84).}
  \item{bio_1}{Annual Mean Temperature (°C).}
  \item{bio_5}{Max Temperature of Warmest Month (°C).}
  \item{bio_6}{Min Temperature of Coldest Month (°C).}
  \item{bio_7}{Temperature Annual Range (bio_5 - bio_6) (°C).}
  \item{bio_12}{Annual Precipitation (mm).}
  \item{bio_13}{Precipitation of Wettest Month (mm).}
  \item{bio_14}{Precipitation of Driest Month (mm).}
  \item{bio_15}{Precipitation Seasonality (Coefficient of Variation).}
}
}
\source{
\url{https://www.worldclim.org}
}
\usage{
back_data
}
\description{
A dataset containing geographic coordinates and two bioclimatic variables
used as a background point cloud for generating and testing niche ellipsoids.
}
\details{
This dataset represents an irregular point cloud typical of environmental
background data used in ecological niche modeling (ENM). It is primarily
used in the `nicheR` package to provide the environmental space for
functions like \code{\link{conserved_ellipses}}.
}
\examples{
data(back_data)
head(back_data)
}
\keyword{datasets}
