% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{BIC.nicher}
\alias{BIC.nicher}
\title{Bayesian information criterion for a nicher fit}
\usage{
\method{BIC}{nicher}(object, ...)
}
\arguments{
\item{object}{a fitted model object for which there exists a
    \code{logLik} method to extract the corresponding log-likelihood, or
    an object inheriting from class \code{logLik}.}

\item{...}{Additional fitted model objects.}
}
\value{
Numeric scalar, or a \code{data.frame} when multiple objects
  are passed.
}
\description{
Computes \eqn{-2 \log L + k \log n}, where \eqn{n} is
\code{nobs(object)}. Calls \code{\link[stats]{BIC}}.
}
