% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ellps.R
\name{build_ellps}
\alias{build_ellps}
\title{Build a 2D or 3D Ellipsoid Object}
\usage{
build_ellps(
  center = c(x = 0, y = 0),
  axes = c(1, 1),
  angles = 0,
  n_points = 100,
  level = 0.95
)
}
\description{
Creates a structured object representing a rotated ellipsoid. This object
contains all necessary matrices for geometric and statistical calculations,
as well as surface points for plotting. This function is a core component for
defining virtual niches in environmental space.
}
