% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_to_tile.R
\name{cell_to_tile}
\alias{cell_to_tile}
\title{Map Tiles Back to Cells from find_niches Output}
\usage{
cell_to_tile(df_final, spe, tile_shift, cores = 10, prob = FALSE)
}
\arguments{
\item{df_final}{A dataframe object from the find_niches output.}

\item{spe}{A SpatialExperiment object.}

\item{cores}{A vector specifying the number of cores to implement (Default: 10).}

\item{prob}{A logical value indicating if the user in interested in calculating the probability (Default: FALSE).}

\item{tile_shiftA}{vector specifying tile shift.}
}
\value{
A list of cells mapped to tiles
}
\description{
Map Tiles Back to Cells from find_niches Output
}
