% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cosmx_means}
\alias{cosmx_means}
\title{cosmx_means}
\format{
An object of class \code{numeric} of length 20310.
}
\source{
{Created to serve as an example.}
}
\usage{
data(cosmx_means)
}
\description{
cosmx_means
}
\examples{
data(cosmx_means)
}
\keyword{datasets}
