% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mask_niche.R
\name{create_mask_niche}
\alias{create_mask_niche}
\title{Crear una máscara para parámetros de modelo de nicho gaussiano}
\usage{
create_mask_niche(p)
}
\arguments{
\item{p}{Número de variables ambientales.}
}
\value{
Vector nombrado con NAs para los parámetros mu y los elementos de la triangular inferior de L.
}
\description{
Crear una máscara para parámetros de modelo de nicho gaussiano
}
