% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_e_space.R
\name{.pe_build_ell2d}
\alias{.pe_build_ell2d}
\title{Helper: build 2D ellipsoid objects for each pair}
\usage{
.pe_build_ell2d(niche)
}
\description{
Helper: build 2D ellipsoid objects for each pair
}
\keyword{internal}
