% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_e_space.R
\name{.pe_get_env_bg}
\alias{.pe_get_env_bg}
\title{Helper: Retrieve environmental background for plot_e_space()}
\usage{
.pe_get_env_bg(env_bg, vs, suitable_env, niche)
}
\description{
Priority:
 1. User-supplied env_bg
 2. From vs via nr_get_env()
 3. If neither exists → message and use suitable_env or niche extents
}
\keyword{internal}
