% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_e_space.R
\name{.pe_get_occ_pts}
\alias{.pe_get_occ_pts}
\title{Helper: Extract occurrence points (df)}
\usage{
.pe_get_occ_pts(occ_pts, vs)
}
\description{
Helper: Extract occurrence points (df)
}
\keyword{internal}
