% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_e_space.R
\name{.pe_get_suitable_pts}
\alias{.pe_get_suitable_pts}
\title{Helper: Get suitable-env points as data.frame for E-space plotting}
\usage{
.pe_get_suitable_pts(suitable_env)
}
\description{
Helper: Get suitable-env points as data.frame for E-space plotting
}
\keyword{internal}
