% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_e_space.R
\name{.pe_make_bbox_df}
\alias{.pe_make_bbox_df}
\title{Helper: Create bounding-box environmental grid from ellipsoid}
\usage{
.pe_make_bbox_df(niche, n = 5000)
}
\description{
Helper: Create bounding-box environmental grid from ellipsoid
}
\keyword{internal}
