% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_e_space.R
\name{.pe_resolve_xyz}
\alias{.pe_resolve_xyz}
\title{Helper: Resolve x,y,z columns}
\usage{
.pe_resolve_xyz(env_bg, x, y, z)
}
\description{
If user did not supply all, infer from env_bg
}
\keyword{internal}
