% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse}
\alias{ellipse}
\title{Ellipse constructor}
\usage{
ellipse(param_list)
}
\arguments{
\item{param_list}{A list of parameters of an ellipse object. First element
is the mu vector. Second element is the square matrix of covariates}
}
\value{
\code{ellipse}
}
\description{
Constructor function that create an ellipse.
}
\examples{
ellipse(get_ellip_par(spOccPnts))

}
