% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_par_example.R
\docType{data}
\name{ellipse_par_example}
\alias{ellipse_par_example}
\title{An ellipse parameter object example. Is a list with two elements.
#' @format A list with two elements. One vector and one matrix:
\describe{
  \item{mu}{mu parameters of the ellipse}
  \item{S}{Covariance parameters of the ellipse}
}}
\format{
An object of class \code{ellipse} of length 2.
}
\usage{
ellipse_par_example
}
\description{
An ellipse parameter object example. Is a list with two elements.
#' @format A list with two elements. One vector and one matrix:
\describe{
  \item{mu}{mu parameters of the ellipse}
  \item{S}{Covariance parameters of the ellipse}
}
}
\keyword{datasets}
