% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_points.R
\name{ellipse_points}
\alias{ellipse_points}
\title{Ellipse points}
\usage{
ellipse_points(par_ellipse, columns = c(1, 2), n_points = 100)
}
\arguments{
\item{par_ellipse}{An ellipse object with ellipse parameters}

\item{columns}{The number of columns of the ellipse to get the points}

\item{n_points}{The number of points for the ellipses to plot}
}
\value{
A data frame with the coordinates to plot
}
\description{
Ellipse points
}
\examples{
ellipse_points(ellipse_par_example)
ellipse_points(ellipse_par_example, columns = c(1, 3)  )
}
