% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_env_m_2d.R
\docType{data}
\name{example_env_m_2d}
\alias{example_env_m_2d}
\title{Samples of environmental data from M hypothesis  to estimate negative log likelihood
from Abeillia abeillei presence points. This is a hummingbird example.
A dataset with two variables containing points
contains 2 bioclimatic variables}
\format{
A data frame with 10000 rows and 2 variables:
\describe{
  \item{bio1WH}{temperature, in C}
  \item{bio12WH}{precipitation, in mm}
}
}
\source{
<doi:10.1016/j.ecolmodel.2021.109823>
}
\usage{
example_env_m_2d
}
\description{
Samples of environmental data from M hypothesis  to estimate negative log likelihood
from Abeillia abeillei presence points. This is a hummingbird example.
A dataset with two variables containing points
contains 2 bioclimatic variables
}
\examples{
head(example_env_m_2d)
}
\keyword{datasets}
