% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_env_m_3d.R
\docType{data}
\name{example_env_m_3d}
\alias{example_env_m_3d}
\title{Samples points from M hypothesis to estimate negative log likelihood
from Abeillia abeillei presence points. This is a
hummingbird example.
A dataset with three columns containing extracted
information from 3 bioclimatic variables.}
\format{
A data frame with 10000 rows and 3 variables:
\describe{
  \item{bio1WH}{Temperature, in C}
  \item{bio12WH}{Precipitation, in mm}
  \item{bio19WH}{Precipitation of Coldest Quarter, in mm}
}
}
\source{
<doi:10.1016/j.ecolmodel.2021.109823>
}
\usage{
example_env_m_3d
}
\description{
Samples points from M hypothesis to estimate negative log likelihood
from Abeillia abeillei presence points. This is a
hummingbird example.
A dataset with three columns containing extracted
information from 3 bioclimatic variables.
}
\keyword{datasets}
