% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_env_occ_3d.R
\docType{data}
\name{example_env_occ_3d}
\alias{example_env_occ_3d}
\title{Species occurrence points to estimate negative log likelihood
from Abeillia abeillei presence points. This is a
hummingbird example.
A dataset with three variables containing points
contains 3 bioclimatic variables}
\format{
A data frame with 73 rows and 3 variables:
\describe{
  \item{bio1WH}{temperature, in C}
  \item{bio12WH}{precipitation, in mm}
  \item{bio19WH}{Precipitation of Coldest Quarter, in mm}
}
}
\source{
<doi:10.1016/j.ecolmodel.2021.109823>
}
\usage{
example_env_occ_3d
}
\description{
Species occurrence points to estimate negative log likelihood
from Abeillia abeillei presence points. This is a
hummingbird example.
A dataset with three variables containing points
contains 3 bioclimatic variables
}
\examples{
head(example_env_occ_3d)
}
\keyword{datasets}
