% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_mu_vec.R
\docType{data}
\name{example_mu_vec}
\alias{example_mu_vec}
\title{Example of a vector of the center of an ellipsoid from two
environmental variables.
vector of length 2 corresponding to the centroid
of an ellipsoid}
\format{
An object of class \code{numeric} of length 2.
}
\source{
<doi:10.1016/j.ecolmodel.2021.109823>
}
\usage{
example_mu_vec
}
\description{
Example of a vector of the center of an ellipsoid from two
environmental variables.
vector of length 2 corresponding to the centroid
of an ellipsoid
}
\examples{
print(example_mu_vec)
}
\keyword{datasets}
