% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_s_mat.R
\docType{data}
\name{example_s_mat}
\alias{example_s_mat}
\title{Example of a covariance matrix of an ellipsoid from two
environmental variables.
The 2 x 2 matrix corresponded to a positive 
semi-definite matrix.
In two dimensions encodes the rotation (orientation) and scaling of an
ellipse.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2 rows and 2 columns.
}
\source{
<doi:10.1016/j.ecolmodel.2021.109823>
}
\usage{
example_s_mat
}
\description{
Example of a covariance matrix of an ellipsoid from two
environmental variables.
The 2 x 2 matrix corresponded to a positive 
semi-definite matrix.
In two dimensions encodes the rotation (orientation) and scaling of an
ellipse.
}
\keyword{datasets}
