% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_niches.R
\name{find_niches}
\alias{find_niches}
\title{Identify Spatial Niches}
\usage{
find_niches(
  spe,
  anno = "Anno",
  barcode = "Barcode",
  x_coord = "x",
  y_coord = "y",
  tile_height = 1000,
  tile_shift = 100,
  image_size = c(4400, 4400),
  num_clusters = 9,
  cores = 10,
  prob = FALSE
)
}
\arguments{
\item{spe}{A SpatialExperiment object. Polygon/x and y spatial co-ordinates, annotation and cell barcode information stored in metadata.}

\item{anno}{A character string specifying the name of the annotation column in spe (Default: "Anno").}

\item{barcode}{A character string specifying the name of the column containing cell barcode information in spe (Default: "Barcode"). Barcode structure: "CellID_fov".}

\item{x_coord}{A character string specifying the name of the column containing x spatial co-ordinates (Default: "x").}

\item{y_coord}{A character string specifying the name of the column containing y spatial co-ordinates (Default: "y").}

\item{tile_height}{A vector specifying tile height (Default: 1000).}

\item{tile_shift}{A vector specifying tile shift (Default: 100).}

\item{image_size}{Vectors specifying image size (Default: c(4400, 4400)).}

\item{num_clusters}{A vector specifying the number of clusters (Default: 9).}

\item{cores}{A vector specifying the number of cores (Default: 10).}

\item{prob}{A logical value indicating if the user in interested in calculating probability (Default: FALSE).}
}
\value{
Dataframe of spatial niches.
}
\description{
Identify Spatial Niches
}
