% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ellipsoid_pars.R
\name{get_ellipsoid_pars}
\alias{get_ellipsoid_pars}
\title{Get ellipsoid parameters. A function to compute average and the inverse of covariance matrix from environmental data}
\usage{
get_ellipsoid_pars(env)
}
\arguments{
\item{env}{A data frame containing environmental variables}
}
\value{
A list with computed average of environmental variables and the
covariance matrix
}
\description{
Get ellipsoid parameters. A function to compute average and the inverse of covariance matrix from environmental data
}
\examples{
get_ellipsoid_pars(example_env_occ_2d)
}
