% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ENM_par.R
\name{get_ENM_par}
\alias{get_ENM_par}
\title{get_ENM_par}
\usage{
get_ENM_par(
  occPts,
  env,
  M_shp = NULL,
  method = c("unbound", "bound", "mahalanobis"),
  fastmode = TRUE,
  itnmax = 100,
  return_best = TRUE
)
}
\arguments{
\item{occPts}{A data frame with ocurrence points. Has species name, longitude and latitude}

\item{env}{A raster stack}

\item{M_shp}{A shape file as M hypothesis}

\item{method}{Method selection. By default, optimization of the negloglike function with lower bound}

\item{fastmode}{logical if is true the L-BFGS-B method is by default}

\item{itnmax}{The maximum number of iterations.}

\item{return_best}{logical. If true return the best optimization. If false return all the possible optimizations}
}
\value{
a list of parameters
}
\description{
get ellipsoid parameters for a Ecollogical fundamental niche model.
}
\examples{
M_path <- system.file("extdata", "Mshp_test.rds", package="nicher")
Mshp <- terra::unwrap(readr::read_rds(M_path))
stack_path <- system.file("extdata", "stack_1_12_crop.rds", package="nicher")
stack_1_12 <- terra::unwrap(readr::read_rds(stack_path))
get_ENM_par(rawSpOccPnts, stack_1_12, Mshp, method = "mahalanobis")
}
