% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_data.R
\name{get_example_data}
\alias{get_example_data}
\title{Function to return mammal shapefiles of Rodentia order}
\usage{
get_example_data(file)
}
\arguments{
\item{file}{An example file for test the r package}
}
\description{
Function to return mammal shapefiles of Rodentia order
}
\examples{
get_example_data(file = "test")
}
