% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{getNicheResource}
\alias{getNicheResource}
\title{retrieve (and cache if necessary) weights or example data for
nicheformer}
\usage{
getNicheResource(entity)
}
\arguments{
\item{entity}{"lungx.zip" (27MB) or "nicheformer_weights.safetensors" (197MB)}
}
\value{
if requesting safetensors, a character path to the cached file; if
requesting lungx.zip,  character path to the cached zip file with the alabaster
serialization of the SCE
}
\description{
retrieve (and cache if necessary) weights or example data for
nicheformer
}
\examples{
zpa = getNicheResource("lungx.zip")
td = tempdir()
unzip(zpa, exdir=td)
scedir = dir(td, patt="lungXen", full=TRUE)
requireNamespace("alabaster.sce")
sce = alabaster.base::readObject(scedir)
sce
}
