% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\value{
The result of calling `rhs(lhs)`.
}
\description{
This re-exports the pipe operator from \code{magrittr} to make it available
for use in your package's functions and for package users.
See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
}
