% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_cells_within_tile.R
\name{identify_cells_within_tile}
\alias{identify_cells_within_tile}
\title{Identify Cells Within Tile (Embedded in find_niches)}
\usage{
identify_cells_within_tile(start_tile, df, tile_height)
}
\arguments{
\item{tile_height}{}
}
\description{
Identify Cells Within Tile (Embedded in find_niches)
}
