% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicher_object.R
\name{is.nicher}
\alias{is.nicher}
\title{Test if an object is of class \code{nicher}}
\usage{
is.nicher(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Logical scalar.
}
\description{
Test if an object is of class \code{nicher}
}
\examples{
is.nicher(42)
}
