% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_niche.R
\name{loglik_niche}
\alias{loglik_niche}
\title{Negative log likelihood of an ellipsoid corrected with environmental
combinations which come from the area of study (M)}
\usage{
loglik_niche(mu, s_mat, env_occ, env_m, neg = TRUE)
}
\arguments{
\item{mu}{A vector mu of parameters}

\item{s_mat}{The covariance matrix from environmental data frame}

\item{env_occ}{A data.frame containing the original sample of environmental
combinations that correspond to presences}

\item{env_m}{A data.frame containing a second random sample of environmental}

\item{neg}{Logical. Default TRUE, returns the negative of the
likelihood}
}
\value{
A negative log likelihood value
}
\description{
Negative log likelihood of an ellipsoid corrected with environmental
combinations which come from the area of study (M)
}
\examples{
loglik_niche(
mu = example_mu_vec,
s_mat = example_s_mat,
env_occ = example_env_occ_2d,
env_m = example_env_m_2d)
#Example with log of the likelihood
loglik_niche(
mu = example_mu_vec,
s_mat = example_s_mat,
env_occ = example_env_occ_2d,
env_m = example_env_m_2d,
neg = FALSE)
}
