% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{ma_bios}
\alias{ma_bios}
\title{Bioclimatic variables for part of the Americas}
\format{
A \code{SpatRaster} with 8 layers:
\describe{
  \item{bio_1}{Annual Mean Temperature}
  \item{bio_5}{Max Temperature of Warmest Month}
  \item{bio_6}{Min Temperature of Coldest Month}
  \item{bio_7}{Temperature Annual Range (bio_5 - bio_6)}
  \item{bio_12}{Annual Precipitation}
  \item{bio_13}{Precipitation of Wettest Month}
  \item{bio_14}{Precipitation of Driest Month}
  \item{bio_15}{Precipitation Seasonality (Coefficient of Variation)}
}
}
\value{
No return value. Used with function \code{\link[terra]{rast}} to
load the GeoTIFF file from the package's \code{inst/extdata} folder.
}
\description{
A \code{SpatRaster} containing 8 bioclimatic variables representing
present-day climatic conditions for an area that covers parts of South
and North America. Variables were obtained at a 10 arc-minute resolution.
Sourced from WorldClim 2.1: \url{https://worldclim.org/data/worldclim21.html}
}
\examples{
ma_bios <- terra::rast(system.file("extdata", "ma_bios.tif",
                                   package = "nicheR"))

terra::plot(ma_bios[[1]])
}
