% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_to_bio_niche.R
\name{math_to_bio_niche}
\alias{math_to_bio_niche}
\title{Convertir parámetros de escala matemática a biológica para modelo de nicho}
\usage{
math_to_bio_niche(param_vector)
}
\arguments{
\item{param_vector}{Vector nombrado con parámetros en escala matemática.
Nombres: mu1..mup, L1..Lk (k = p*(p+1)/2).
Las diagonales de L deben estar en escala logarítmica.}
}
\value{
Lista con elementos `mu` (vector) y `L` (matriz triangular inferior).
}
\description{
Convertir parámetros de escala matemática a biológica para modelo de nicho
}
