% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_comunities.R
\name{nested_ellipses}
\alias{nested_ellipses}
\title{Generate nested ellipses based on a reference ellipse}
\usage{
nested_ellipses(object, n = 10, smallest_proportion = 0.1, bias = 1)
}
\arguments{
\item{object}{An object of class "nicheR_ellipsoid"
describing an initial ellipse. Must contain \code{centroid},
\code{cov_matrix}, and \code{cl}.}

\item{n}{Integer. Number of nested ellipses to generate. Default is 10.}

\item{smallest_proportion}{Numeric scalar in \code{(0, 1)}. The scale of the
smallest ellipse relative to the original. Default is \code{0.1}.}

\item{bias}{Numeric. An exponent controlling the spacing of the nested
ellipses.
\itemize{
  \item \code{bias = 1}: Linear spacing (default).
  \item \code{0 < bias < 1}: Clusters ellipses toward the **border**
  (outer original ellipse).
  \item \code{bias > 1}: Clusters ellipses toward the **centroid**
  (inner smallest ellipse).
}}
}
\value{
An object of class \code{nicheR_community} containing the generated
ellipses, the reference object, and generation metadata.
}
\description{
Creates a sequence of nested ellipses by scaling the covariance matrix of a
reference ellipse. The distribution of the nested ellipses can be controlled
using a bias exponent to cluster them toward the border or the centroid.
}
\details{
The largest ellipse corresponds to the original reference, and the
smallest is that scaled by \code{smallest_proportion}.

The function generates a sequence of scale factors $k$ using the formula:
$k_i = \text{smaller\_proportion} + (1 - \text{smaller\_proportion}) \ times t_i^{\text{bias}}$,
where $t_i$ is a linear sequence from 1 down to 0.#' A list of \code{n} "nicheR_ellipsoid" objects.
}
