% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{nicheformer_load_weights}
\alias{nicheformer_load_weights}
\title{Load pretrained Nicheformer weights from a safetensors file}
\usage{
nicheformer_load_weights(model, path)
}
\arguments{
\item{model}{A \code{nicheformer} model instance}

\item{path}{Path to the safetensors weights file}
}
\value{
The model with loaded weights, set to eval mode
}
\description{
Load pretrained Nicheformer weights from a safetensors file
}
\examples{
nn = nicheformer()
nn$state_dict()[["token_emb.weight"]]$std()$item() # S.D. of all elements in tensor
if (nchar(wpa <- Sys.getenv("LOCAL_NICHE_WEIGHTS"))>0) {
nicheformer_load_weights(nn, wpa)  # will print
nn$state_dict()[["token_emb.weight"]]$std()$item()  # should be smaller than previous
}
}
