% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{nicheformer_weights}
\alias{nicheformer_weights}
\title{Download and cache the Nicheformer weights}
\usage{
nicheformer_weights()
}
\value{
path to the cached .pt file
}
\description{
Download and cache the Nicheformer weights
}
