% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ellpisoid.R
\name{parse_cor_tilt}
\alias{parse_cor_tilt}
\title{Parse Correlation Tilt Matrix}
\usage{
parse_cor_tilt(cor_tilt, p)
}
\arguments{
\item{cor_tilt}{Correlation structure describing variable co-variation.
Can be:
\itemize{
\item \code{NULL} for no tilt (identity matrix)
\item a \eqn{p \times p} correlation matrix
\item a matrix or data.frame of observations used to estimate correlation
}}

\item{p}{Integer. Number of environmental dimensions.}
}
\value{
A \eqn{p \times p} correlation matrix.
}
\description{
Converts user-supplied correlation information into a valid correlation
matrix controlling ellipsoid orientation in environmental space.
}
\details{
The correlation matrix defines ellipsoid orientation but does not affect
marginal niche breadth, which is controlled separately by standard deviations.
}
\keyword{internal}
