% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ellipse.R
\name{plot.ellipse}
\alias{plot.ellipse}
\title{Plot method for "ellipse" class}
\usage{
\method{plot}{ellipse}(x, ...)
}
\arguments{
\item{x}{An object of class "ellipse"}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Plot method for "ellipse" class
}
