% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_e_space.R
\name{plot_e_space}
\alias{plot_e_space}
\title{Plot niche and environments in E-space (2D or 3D)}
\usage{
plot_e_space(
  env_bg = NULL,
  x,
  y,
  z,
  labels = c("ENV 1", "ENV 2", "ENV 3"),
  n_bg = 10000,
  niche = NULL,
  suitable_env = NULL,
  occ_pts = NULL,
  rand_seed = 1234,
  show.occ.density = FALSE,
  plot.3d = FALSE,
  colors = NULL,
  palette = "default",
  vs = NULL
)
}
\description{
Automatically retrieves:
 - env_bg (environmental background),
 - suitable_env (inside ellipsoid),
 - occurrences,
 - x,y,z predictors
from: user args → vs → nr_get().
}
