% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities_3d.R
\name{plot_ellipsoid_3d}
\alias{plot_ellipsoid_3d}
\title{Plot a nicheR ellipsoid in 3D environmental space}
\usage{
plot_ellipsoid_3d(object, dim = c(1, 2, 3), wire = FALSE, aspect = TRUE,
                  background = NULL, prediction = NULL, col_layer = NULL,
                  pal = hcl.colors(100, palette = "Viridis"),
                  rev_pal = FALSE, bg_sample = NULL, col_ell = "#8b0000",
                  alpha_ell = 1, alpha_bg = 1, col_bg = "#8A8A8A",
                  fixed_lims = NULL, xlab = NULL, ylab = NULL,
                  zlab = NULL, ...)
}
\arguments{
\item{object}{A \code{nicheR_ellipsoid} object constructed with at least
3 dimensions.}

\item{dim}{Integer vector of length 3. Indices of dimensions to plot.}

\item{wire}{Logical. If \code{TRUE}, plots wireframe. The default,
\code{FALSE}, plots a shaded volume.}

\item{aspect}{Logical. If \code{TRUE}, maintains aspect ratio (1:1:1).}

\item{background}{Optional data frame/matrix of background points.
This argument has priority over \code{prediction} if both are provided.}

\item{prediction}{Optional data frame/matrix for prediction surfaces.}

\item{col_layer}{Character or \code{NULL}. Column in \code{prediction}
to use for coloring.}

\item{pal}{Color palette function or character vector.}

\item{rev_pal}{Logical. If \code{TRUE}, reverses the palette.}

\item{bg_sample}{Integer or \code{NULL}. Subsample size for large data.}

\item{col_ell}{Color of the ellipsoid. Default is \code{"#000000"}.}

\item{alpha_ell}{Transparency of the ellipsoid. Default is \code{1}.
Not applied to wireframe mode.}

\item{alpha_bg}{Transparency of background points. Default is \code{1}.
Also applied to prediction points.}

\item{col_bg}{Color for background points.}

\item{fixed_lims}{Named list with \code{xlim}, \code{ylim}, and \code{zlim}.}

\item{xlab}{x-axis label. The default, \code{NULL}, uses ellipsoid object
variable names, if any found.}

\item{ylab}{y-axis label. The default, \code{NULL}, uses ellipsoid object
variable names, if any found.}

\item{zlab}{z-axis label. The default, \code{NULL}, uses ellipsoid object
variable names, if any found.}

\item{...}{Additional graphical parameters.}
}
\value{
A 3d plot of the ellipsoid in E-space, shown in viewer.
}
\description{
Creates an interactive 3D plot of a \code{nicheR_ellipsoid} object with
support for background points or suitability surfaces.
}
\examples{
# Building an ellipsoid
## Define ranges for three variables
range <- data.frame(bio_1  = c(22, 32),
                    bio_12 = c(800, 4200),
                    bio_15 = c(45, 115))

## Build the ellipsoid
ell5 <- build_ellipsoid(ranges = range)
ell5$cov_limits

ell5 <- update_ellipsoid_covariance(ell5, c("bio_1-bio_12" = 200,
                                            "bio_1-bio_15" = 0,
                                            "bio_12-bio_15" = -3000))

# Plot the ellipsoid in 3D
\donttest{
if(requireNamespace("rgl", quietly = TRUE)){
  plot_ellipsoid_3d(ell5)
 }
}
}
