% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities.R
\name{plot_nicheR}
\alias{plot_nicheR}
\title{Plot 2d base R just for testing right now}
\usage{
plot_nicheR(
  ell_list = NULL,
  dims = c(1, 2),
  n_points = 200,
  labels = NULL,
  xlab = NULL,
  ylab = NULL,
  main = "Ellipsoid",
  add = FALSE,
  legend_pos = "topright",
  occ_list = NULL,
  occ_pch = 1,
  occ_col = "green4",
  occ_cex = 0.6,
  ...
)
}
\description{
Plot 2d base R just for testing right now
}
