% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ellipse.R
\name{predict.ellipse}
\alias{predict.ellipse}
\title{Predict method for "ellipse" class}
\usage{
\method{predict}{ellipse}(object, layers, returnTable = FALSE, ...)
}
\arguments{
\item{object}{An object of class "ellipse"}

\item{layers}{An object of class "rast" from terra package where to evaluate the ellipse model}

\item{returnTable}{logical if true return table instead a raster with the prediction}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Predict method for "ellipse" class
}
