% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ellipse.R
\name{print.ellipse}
\alias{print.ellipse}
\title{Print method for "ellipse" class}
\usage{
\method{print}{ellipse}(x, ...)
}
\arguments{
\item{x}{An object of class "ellipse"}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Print method for "ellipse" class
}
