% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_virtual_species.R
\name{print.NicheR_species}
\alias{print.NicheR_species}
\title{Print a NicheR virtual species}
\usage{
\method{print}{NicheR_species}(x, ...)
}
\arguments{
\item{x}{A \code{NicheR_species} object.}

\item{...}{Not used.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a NicheR virtual species
}
