% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refsce}
\alias{refsce}
\title{refsce}
\format{
An object of class \code{SingleCellExperiment} with 20310 rows and 1 columns.
}
\source{
{Created to serve as an example.}
}
\usage{
data(refsce)
}
\description{
refsce
}
\examples{
data(refsce)
}
\keyword{datasets}
