% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid_utilities.R
\name{resolve_prediction}
\alias{resolve_prediction}
\title{Resolves user input when to many columns or layers}
\usage{
resolve_prediction(prediction, prediction_layer)
}
\description{
Resolves user input when to many columns or layers
}
