% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samMPts_1_12_19.R
\docType{data}
\name{samMPts_1_12_19}
\alias{samMPts_1_12_19}
\title{Samples points from M hypothesis  to estimate negative log likelihood
from Abeillia abeillei presence points. This is a
hummingbird example.
A dataset with two variables containing points
contains 2 bioclimatic variables
#' @format A data frame with 10000 rows and 2 variables:
\describe{
  \item{bio1WH}{Yemperature, in C}
  \item{bio12WH}{Precipitaion, in mm}
  \item{bio19WH}{Precipitation of Coldest Quarter, in mm}
}}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 10000 rows and 3 columns.
}
\source{
<doi:10.1016/j.ecolmodel.2021.109823>
}
\usage{
samMPts_1_12_19
}
\description{
Samples points from M hypothesis  to estimate negative log likelihood
from Abeillia abeillei presence points. This is a
hummingbird example.
A dataset with two variables containing points
contains 2 bioclimatic variables
#' @format A data frame with 10000 rows and 2 variables:
\describe{
  \item{bio1WH}{Yemperature, in C}
  \item{bio12WH}{Precipitaion, in mm}
  \item{bio19WH}{Precipitation of Coldest Quarter, in mm}
}
}
\keyword{datasets}
