% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{sinusoidal_pe}
\alias{sinusoidal_pe}
\title{Sinusoidal positional encoding (used when learnable_pe = FALSE)}
\usage{
sinusoidal_pe(context_length, dim_model)
}
\arguments{
\item{context_length}{numeric}

\item{dim_model}{numeric}
}
\description{
Sinusoidal positional encoding (used when learnable_pe = FALSE)
}
