% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spOccPnts_1_12_19.R
\docType{data}
\name{spOccPnts_1_12_19}
\alias{spOccPnts_1_12_19}
\title{Species ocurrence points to estimate negative log likelihood
from Abeillia abeillei presence points. This is a
hummingbird example.
A dataset with two variables containing points
contains 2 bioclimatic variables
#' @format A data frame with 73 rows and 2 variables:
\describe{
  \item{bio1WH}{temperature, in C}
  \item{bio12WH}{precipitaion, in mm}
  \item{bio19WH}{Precipitation of Coldest Quarter, in mm}
}}
\format{
An object of class \code{data.frame} with 73 rows and 3 columns.
}
\source{
<doi:10.1016/j.ecolmodel.2021.109823>
}
\usage{
spOccPnts_1_12_19
}
\description{
Species ocurrence points to estimate negative log likelihood
from Abeillia abeillei presence points. This is a
hummingbird example.
A dataset with two variables containing points
contains 2 bioclimatic variables
#' @format A data frame with 73 rows and 2 variables:
\describe{
  \item{bio1WH}{temperature, in C}
  \item{bio12WH}{precipitaion, in mm}
  \item{bio19WH}{Precipitation of Coldest Quarter, in mm}
}
}
\keyword{datasets}
