% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xenium_means}
\alias{xenium_means}
\title{xenium_means}
\format{
An object of class \code{numeric} of length 20310.
}
\source{
{Created to serve as an example.}
}
\usage{
data(xenium_means)
}
\description{
xenium_means
}
\examples{
data(xenium_means)
}
\keyword{datasets}
