% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities_3d.R
\name{add_data_3d}
\alias{add_data_3d}
\title{Add data to an existing 3D E-space plot}
\usage{
add_data_3d(data, dim = c(1, 2, 3), col_layer = NULL, ...)
}
\arguments{
\item{data}{A data frame or matrix containing the points.}

\item{dim}{Integer vector of length 3. Indices of dimensions to plot.}

\item{col_layer}{Character or \code{NULL}. Column for coloring.}

\item{...}{Additional arguments passed to \code{rgl::points3d}.}
}
\description{
Add data to an existing 3D E-space plot
}
