% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities_3d.R
\name{add_data_3d}
\alias{add_data_3d}
\title{Add data to an existing 3D E-space plot}
\usage{
add_data_3d(data, dim = c(1, 2, 3), col_layer = NULL, alpha = 1, ...)
}
\arguments{
\item{data}{A data frame or matrix containing the points.}

\item{dim}{Integer vector of length 3. Indices of dimensions to plot.}

\item{col_layer}{Character or \code{NULL}. Column for coloring.}

\item{alpha}{Transparency for the points. Default is \code{1}.}

\item{...}{Additional arguments passed to \code{rgl::points3d}.}
}
\description{
Add data to an existing 3D E-space plot
}
\examples{
# Building an ellipsoid
## Define ranges for three variables
range <- data.frame(bio_1  = c(22, 32),
                    bio_12 = c(800, 4200),
                    bio_15 = c(45, 115))

## Build the ellipsoid
ell5 <- build_ellipsoid(range = range)
ell5$cov_limits

ell5u <- update_ellipsoid_covariance(ell5, c("bio_1-bio_12" = 200,
                                             "bio_1-bio_15" = 0,
                                             "bio_12-bio_15" = -3000))
\dontrun{
# Plot the ellipsoid in 3D
plot_ellipsoid_3d(ell5u)

# Add background points
add_data_3d(back_data[, c(3, 7, 10)], col = "#8A8A8A")
}
}
