% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities_3d.R
\name{add_ellipsoid_3d}
\alias{add_ellipsoid_3d}
\title{Add an ellipsoid to an existing 3D E-space plot}
\usage{
add_ellipsoid_3d(
  object,
  dim = c(1, 2, 3),
  wire = FALSE,
  col_ell = "#800000",
  alpha_ell = 1,
  ...
)
}
\arguments{
\item{object}{A \code{nicheR_ellipsoid} object.}

\item{dim}{Integer vector of length 3.}

\item{wire}{Logical. If \code{TRUE}, plots wireframe, otherwise plots a
shaded volume. Default is \code{FALSE}.}

\item{col_ell}{Color of the ellipsoid. Default is \code{"#000000"}.}

\item{alpha_ell}{Transparency of the ellipsoid. Default is
\code{1}. Not applied if \code{wire = TRUE}.}

\item{...}{Additional arguments passed to \code{rgl::wire3d} or
\code{rgl::shade3d}.}
}
\description{
Add an ellipsoid to an existing 3D E-space plot
}
