% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utilities_3d.R
\name{add_ellipsoid_3d}
\alias{add_ellipsoid_3d}
\title{Add an ellipsoid to an existing 3D E-space plot}
\usage{
add_ellipsoid_3d(object, dim = c(1, 2, 3), wire = FALSE, col_ell = "#8b0000",
                  alpha_ell = 1, ...)
}
\arguments{
\item{object}{A \code{nicheR_ellipsoid} object.}

\item{dim}{Integer vector of length 3.}

\item{wire}{Logical. If \code{TRUE}, plots wireframe, otherwise plots a
shaded volume. Default is \code{FALSE}.}

\item{col_ell}{Color of the ellipsoid. Default is \code{"#000000"}.}

\item{alpha_ell}{Transparency of the ellipsoid. Default is
\code{1}. Not applied if \code{wire = TRUE}.}

\item{...}{Additional arguments passed to \code{rgl::wire3d} or
\code{rgl::shade3d}.}
}
\description{
Add an ellipsoid to an existing 3D E-space plot
}
\examples{
# Building an ellipsoid
## Define ranges for three variables
range <- data.frame(bio_1  = c(22, 32),
                    bio_12 = c(800, 4200),
                    bio_15 = c(45, 115))

## Build the ellipsoid
ell5 <- build_ellipsoid(range = range)
ell5$cov_limits

ell5u <- update_ellipsoid_covariance(ell5, c("bio_1-bio_12" = 200,
                                             "bio_1-bio_15" = 0,
                                             "bio_12-bio_15" = -3000))

# Plot the ellipsoid in 3D
plot_ellipsoid_3d(ell5u)

# Add the original ellipsoid as a wireframe
add_ellipsoid_3d(ell5, wire = TRUE, col = "#0e008b")
}
