% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_to_tile.R
\name{cell_to_tile}
\alias{cell_to_tile}
\title{Map Tiles Back to Cells from find_niches Output}
\usage{
cell_to_tile(
  df_final,
  spe,
  fov = "fov",
  barcode = "Barcode",
  data_type = c("polygon", "dot"),
  x_coord = "x",
  y_coord = "y",
  x_centre = "CenterX_local_px",
  y_centre = "CenterY_local_px",
  tile_shift,
  cores = 10,
  prob = FALSE
)
}
\arguments{
\item{df_final}{A dataframe object from find_niches output.}

\item{spe}{A SpatialExperiment object. Polygon/x and y spatial co-ordinates, annotation and cell barcode information stored in metadata.}

\item{fov}{A character string specifying the name of the fields of view column in spe (Default: 'fov').}

\item{barcode}{A character string specifying the name of the column containing cell barcode information in spe (Default: 'Barcode'). Barcode structure: 'CellID_fov'.}

\item{data_type}{A character string indicating if spe contains polygon or dot/point information (Options: 'polygon' or 'dot').}

\item{x_coord}{A character string specifying the name of the column containing x spatial co-ordinates (Default: 'x'). If fov=NULL these will be global coordinates, otherwise, local coordinates.}

\item{y_coord}{A character string specifying the name of the column containing y spatial co-ordinates (Default: 'y'). If fov=NULL these will be global coordinates, otherwise, local coordinates.}

\item{x_centre}{A character string specifying the name of the column containing x centroid co-ordinates (Default: 'CenterX_local_px'). Applicable for data_type='polygon'.}

\item{y_centre}{A character string specifying the name of the column containing y centroid co-ordinates (Default: 'CenterY_local_px'). Applicable for data_type='polygon'.}

\item{tile_shift}{A vector specifying tile shift.}

\item{cores}{A vector specifying the number of cores to implement (Default: 10).}

\item{prob}{A logical value indicating if the user in interested in calculating the probability (Default: FALSE).}
}
\value{
A list of cells mapped to tiles
}
\description{
Map Tiles Back to Cells from find_niches Output
}
