% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_virtual_species.R
\name{create_virtual_species}
\alias{create_virtual_species}
\title{Create a virtual species end to end
Routes ... to build_ellps(), get_suitable_env(), and get_sample_occ()
Returns an S3 object of class "NicheR_species"}
\usage{
create_virtual_species(
  ...,
  out.file = FALSE,
  out.file.name = NULL,
  verbose = TRUE
)
}
\description{
Create a virtual species end to end
Routes ... to build_ellps(), get_suitable_env(), and get_sample_occ()
Returns an S3 object of class "NicheR_species"
}
