% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_niche.R
\name{.optimize_niche_helper_cpp}
\alias{.optimize_niche_helper_cpp}
\title{Run ucminfcpp::ucminf_xptr() for a single starting vector.}
\usage{
.optimize_niche_helper_cpp(
  param,
  env_occ,
  env_m,
  control,
  likelihood,
  grad,
  weighted_inputs = NULL,
  ...
)
}
\description{
Run ucminfcpp::ucminf_xptr() for a single starting vector.
}
\keyword{internal}
