% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid_calculator.R
\name{ellipsoid_calculator}
\alias{ellipsoid_calculator}
\title{Calculate n-dimensional ellipsoid metrics}
\usage{
ellipsoid_calculator(cov_matrix, centroid, cl, verbose = TRUE)
}
\arguments{
\item{cov_matrix}{A square, numeric covariance matrix representing the
shape and orientation of the niche.}

\item{centroid}{A numeric vector representing the center of the ellipsoid
(e.g., the mean of the variables).}

\item{cl}{A numeric value between 0 and 1 representing the confidence
level (e.g., 0.95 for a 95\% confidence ellipsoid).}

\item{verbose}{Logical; if \code{TRUE}, the function prints progress messages
to the console. Default is \code{TRUE}.}
}
\value{
An object of class \code{\link{nicheR_ellipsoid}}. With all elements
computed for the ellipsoid produced by the input parameters.
}
\description{
This function computes the geometric and statistical characteristics of an
n-dimensional ellipsoid based on a covariance matrix and a centroid.
It calculates the semi-axes lengths, the coordinates of the vertices,
the volume, and checks for Spectral Positive Definiteness (SPD).
}
\examples{
cm <- matrix(c(11.11, 0, 0, 17777.78), nrow = 2)
ctr <- c(20, 600)
ellie <- ellipsoid_calculator(cm, ctr, 0.95)
}
