% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_occ_df.R
\docType{data}
\name{example_occ_df}
\alias{example_occ_df}
\title{Species occurrence points from Abeillia abeillei presence points after
download and clean from GBIF. This is a hummingbird example.
A dataset with three variables. Contains scientific name, longitude and
latitude.}
\format{
A data frame with 73 rows and 3 variables:
\describe{
  \item{species}{Species name}
  \item{long}{Decimal longitude geographical coordinate, in degrees}
  \item{lat}{Decimal latitude geographical coordinate, in degrees}
}
}
\source{
<doi:10.1016/j.ecolmodel.2021.109823>
}
\usage{
example_occ_df
}
\description{
Species occurrence points from Abeillia abeillei presence points after
download and clean from GBIF. This is a hummingbird example.
A dataset with three variables. Contains scientific name, longitude and
latitude.
}
\keyword{datasets}
