% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_niches.R
\name{find_niches}
\alias{find_niches}
\title{Identify Spatial Niches}
\usage{
find_niches(
  spe,
  fov = "fov",
  anno = "Anno",
  barcode = "Barcode",
  data_type = c("polygon", "dot"),
  x_coord = "x_local_px",
  y_coord = "y_local_px",
  x_centre = "CenterX_local_px",
  y_centre = "CenterY_local_px",
  tile_height = 1000,
  tile_shift = 100,
  image_size = c(4400, 4400),
  num_clusters = 9,
  cores = 10,
  prob = FALSE
)
}
\arguments{
\item{spe}{A SpatialExperiment object. Polygon/x and y spatial co-ordinates, annotation and cell barcode information stored in metadata.}

\item{fov}{A character string specifying the name of the fields of view column in spe (Default: 'fov').}

\item{anno}{A character string specifying the name of the annotation column in spe (Default: 'Anno').}

\item{barcode}{A character string specifying the name of the column containing cell barcode information in spe (Default: 'Barcode'). Barcode structure: 'CellID_fov'.}

\item{data_type}{A character string indicating if spe contains polygon or dot/point information (Options: 'polygon' or 'dot').}

\item{x_coord}{A character string specifying the name of the column containing x spatial co-ordinates (Default: 'x_local_px').}

\item{y_coord}{A character string specifying the name of the column containing y spatial co-ordinates (Default: 'y_local_px').}

\item{x_centre}{A character string specifying the name of the column containing x centroid co-ordinates (Default: 'CenterX_local_px'). Applicable for data_type='polygon'.}

\item{y_centre}{A character string specifying the name of the column containing y centroid co-ordinates (Default: 'CenterY_local_px'). Applicable for data_type='polygon'.}

\item{tile_height}{A vector specifying tile height (Default: 1000).}

\item{tile_shift}{A vector specifying tile shift (Default: 100).}

\item{image_size}{Vectors specifying image size (Default: c(4400, 4400)).}

\item{num_clusters}{A vector specifying the number of clusters (Default: 9).}

\item{cores}{A vector specifying the number of cores (Default: 10).}

\item{prob}{A logical value indicating if the user in interested in calculating probability (Default: FALSE).}
}
\value{
Dataframe of spatial niches.
}
\description{
Identify Spatial Niches
}
