% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_niches.R
\name{find_niches}
\alias{find_niches}
\title{Find Spatial Niches}
\usage{
find_niches(
  df_polygons,
  spe,
  anno = "Anno",
  barcode = "Barcode",
  tile_height = 1000,
  tile_shift = 100,
  image_size = c(4400, 4400),
  num_clusters = 9,
  cores = 10,
  prob = FALSE
)
}
\arguments{
\item{df_polygons}{A dataframe object containing spatial polygon information.}

\item{spe}{A SpatialExperiment object.}

\item{anno}{A character string specifying the name of the annotation column in spe (Default: "Anno").}

\item{barcode}{A character string specifying the name of the column containing cell barcode information in spe (Default: "Barcode").}

\item{tile_height}{A vector specifying tile height (Default: 1000).}

\item{tile_shift}{A vector specifying tile shift (Default: 100).}

\item{image_size}{Vectors specifying image size (Default: c(4400, 4400)).}

\item{num_clusters}{A vector specifying the number of clusters (Default: 9).}

\item{cores}{A vector specifying the number of cores to implement (Default: 10).}

\item{prob}{A logical value indicating if the user in interested in calculating the probability (Default: FALSE).}
}
\value{
A dataframe of niches
}
\description{
Find Spatial Niches
}
