% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_range_df_niche.R
\name{get_range_df_niche}
\alias{get_range_df_niche}
\title{Generate reasonable ranges for starting parameters (math scale)}
\usage{
get_range_df_niche(env_data, quant_vec = c(0.1, 0.5, 0.9))
}
\arguments{
\item{env_data}{Data frame of environmental values.}

\item{quant_vec}{Numeric vector of length 3: lower, middle, upper quantiles.}
}
\value{
A data frame with rows corresponding to parameter names and columns
lower, center, upper.
}
\description{
Generate reasonable ranges for starting parameters (math scale)
}
\keyword{internal}
