% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_suitable_env.R
\name{get_suitable_env}
\alias{get_suitable_env}
\title{Extract Suitable Environmental Area from a Niche Ellipsoid}
\usage{
get_suitable_env(
  niche,
  env_bg,
  out.suit = c("data.frame", "spatial", "both"),
  distances = FALSE,
  level = 1,
  verbose = TRUE
)
}
\description{
This function identifies and extracts environmental grid cells or data
points that fall within a defined ellipsoid niche based on Mahalanobis
distance.
}
