% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_bio_niche.R
\name{loglik_bio_niche}
\alias{loglik_bio_niche}
\title{Log-verosimilitud para modelo de nicho gaussiano (escala biológica)}
\usage{
loglik_bio_niche(sam1, sam2, mu, L)
}
\arguments{
\item{sam1}{Data.frame o matriz de puntos de presencia (filas = observaciones, columnas = variables).}

\item{sam2}{Data.frame o matriz de puntos de fondo ambiental.}

\item{mu}{Vector de medias (óptimo) de longitud p.}

\item{L}{Matriz triangular inferior de la descomposición de Cholesky de la matriz de covarianza S (S = L %*% t(L)).
Las diagonales deben ser positivas.}
}
\value{
Log-verosimilitud (escalar).
}
\description{
Log-verosimilitud para modelo de nicho gaussiano (escala biológica)
}
