% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_nicheR_ellipsoid}
\alias{new_nicheR_ellipsoid}
\title{nicheR_ellipsoid Class Constructor}
\usage{
new_nicheR_ellipsoid(
  dimensions,
  var_names,
  centroid,
  cov_matrix,
  Sigma_inv,
  chol_Sigma,
  eigen,
  cl,
  chi2_cutoff,
  semi_axes_lengths,
  axes_coordinates,
  volume,
  cov_limits
)
}
\arguments{
\item{dimensions}{The number of variables/dimensions.}

\item{var_names}{Names of the original variables.}

\item{centroid}{The numeric center vector.}

\item{cov_matrix}{The covariance matrix.}

\item{Sigma_inv}{The precision matrix.}

\item{chol_Sigma}{The Cholesky decomposition.}

\item{eigen}{List of eigenvectors and eigenvalues.}

\item{cl}{Confidence level.}

\item{chi2_cutoff}{Chi-squared quantile.}

\item{semi_axes_lengths}{Radii of the ellipsoid.}

\item{axes_coordinates}{List of vertex matrices.}

\item{volume}{Calculated hyper-volume.}

\item{cov_limits}{Axis-aligned limits.}
}
\value{
An object of class \code{nicheR_ellipsoid} with the fields described above.
}
\description{
Internal helper to create a nicheR_ellipsoid object.
}
\keyword{internal}
