% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{nicheformer}
\alias{nicheformer}
\title{the nicheformer interface}
\usage{
nicheformer(
  dim_model = 512L,
  nheads = 16L,
  dim_feedforward = 1024L,
  nlayers = 12L,
  dropout = 0,
  masking_p = 0.15,
  n_tokens = 20340L,
  total_tokens = 20345L,
  context_length = 1500L,
  learnable_pe = TRUE
)
}
\arguments{
\item{dim_model}{integer}

\item{nheads}{integer}

\item{dim_feedforward}{integer}

\item{nlayers}{integer}

\item{dropout}{numeric}

\item{masking_p}{numeric}

\item{n_tokens}{integer}

\item{total_tokens}{integer}

\item{context_length}{integer}

\item{learnable_pe}{logical}
}
\description{
the nicheformer interface
}
