% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicheR-package.R
\docType{package}
\name{nicheR-package}
\alias{nicheR}
\alias{nicheR-package}
\title{nicheR: Ellipsoid-Based Virtual Niches and Visualization}
\description{
Provides a robust set of tools for researchers and modelers to construct and define virtual ecological niches using ellipsoid geometries. It enables the identification and extraction of suitable environmental areas, simulation of species occurrence points with various sampling strategies, and visualization of niche boundaries and simulated occurrences in both environmental and geographic space. Inspired by methodologies in 'NicheA' and the 'virtualspecies' R package, 'nicheR' aims to streamline the process of niche conceptualization and data generation for ecological studies. Methodological and theoretical foundations are described in Peterson et al. (2011, ISBN:9780691136882), Etherington et al. (2009) \doi{10.1111/j.1365-2699.2008.02041.x}, Qiao et al. (2015) \doi{10.1111/ecog.01961}, Nunez-Penichet et al. (2021) \doi{10.21425/F5FBG52142}, Cobos and Peterson (2022) \doi{10.17161/bi.v17i.15985}, Alkishe et al. (2022) \doi{10.5194/we-22-33-2022}, and Leroy et al. (2015) \doi{10.1111/ecog.01388}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/castanedaM/nicheR}
  \item \url{https://castanedam.github.io/nicheR/}
  \item Report bugs at \url{https://github.com/castanedaM/nicheR/issues}
}

}
\author{
\strong{Maintainer}: Mariana Castaneda-Guzman \email{marianacg@vt.edu} (\href{https://orcid.org/0000-0001-6106-4284}{ORCID})

Authors:
\itemize{
  \item Mariana Castaneda-Guzman \email{marianacg@vt.edu} (\href{https://orcid.org/0000-0001-6106-4284}{ORCID})
  \item Connor Hughes \email{connorjh@vt.edu} (\href{https://orcid.org/0000-0002-3720-0837}{ORCID})
  \item Paanwaris Paansri \email{paanwaris@vt.edu} (\href{https://orcid.org/0000-0001-9992-098X}{ORCID})
  \item Marlon E. Cobos \email{manubio13@gmail.com} (\href{https://orcid.org/0000-0002-2611-1767}{ORCID})
}

}
\keyword{internal}
