% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicheR-package.R
\docType{package}
\name{nicheR}
\alias{nicheR}
\alias{nicheR-package}
\title{Method to identify niches of co-occuring cell types from spatial
transcriptomics and proteomics data with single cell resolution.}
\description{
`nicheR` implements a convolutional network type approach to identify niches
 in spatial omics data with single cell resolution. `nicheR` works with the
 \link{SpatialExperiment::SpatialExperiment} class and requires cell type
 annotations as inputs.
}
\details{
Key neighborhood analysis functions include \code{\link{find_niches}}, \code{\link{cell_to_tile}}
}
\author{
Montana Spiteri \email{spiteri.m@wehi.edu.au}
}
\keyword{internal}
