% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ellipsoid.R
\name{plot_ellipsoid}
\alias{plot_ellipsoid}
\title{Plot a nicheR Ellipsoid}
\usage{
plot_ellipsoid(
  object,
  background = NULL,
  dim = c(1, 2),
  bg_sample = 1000,
  lty = 1,
  lwd = 1,
  col_ell = "#000000",
  col_bg = "#8A8A8A",
  pch = 1,
  alpha_bg = 0.5,
  alpha_ell = 0.7,
  cex_ell = 1,
  cex_bg = 1,
  ...
)
}
\arguments{
\item{object}{A nicheR ellipsoid object containing \code{centroid},
\code{cov_matrix}, and \code{chi2_cutoff}.}

\item{background}{Optional data frame of background points (rows = observations,
columns = environmental variables).}

\item{bg_sample}{Integer. Number of background points to bg_sample for plotting.}

\item{lty}{Line type for ellipsoid boundary.}

\item{lwd}{Line width for ellipsoid boundary.}

\item{col_ell}{Color for ellipsoid boundary.}

\item{col_bg}{Color for background points.}

\item{pch}{Plotting symbol for background points.}

\item{alpha_bg}{Transparency for background points.}

\item{alpha_ell}{Transparency for ellipsoid boundary.}

\item{cex_ell}{Expansion factor for ellipsoid line.}

\item{cex_bg}{Expansion factor for background points.}

\item{...}{Additional graphical parameters passed to \code{plot()}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Plots the boundary of a probabilistic ellipsoid in environmental space,
optionally overlaying background points.
}
