% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_bias.R
\name{prepare_bias}
\alias{prepare_bias}
\title{Prepare Sampling Bias Surfaces}
\usage{
prepare_bias(
  bias_surface,
  effect_direction = c("direct", "inverse"),
  template_layer = NULL,
  include_composite = TRUE,
  include_processed_layers = FALSE,
  mask_na = FALSE,
  verbose = TRUE
)
}
\description{
Prepare Sampling Bias Surfaces
}
