% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print}
\alias{print}
\alias{print.nicheR_ellipsoid}
\alias{print,nicheR_nicheR_ellipsoid-method}
\alias{print,nicheR_nicheR_community-method}
\alias{print.nicheR_community}
\title{Print method for nicheR objects}
\usage{
\method{print}{nicheR_ellipsoid}(x, digits = 3, ...)

\method{print}{nicheR_community}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of the classes \code{"nicheR_ellipsoid"} or
\code{"nicheR_community"}.}

\item{digits}{Integer. Number of decimal places used when printing numeric
values. Default is 3.}

\item{...}{Additional arguments.}
}
\value{
The input object \code{x}, returned invisibly.
}
\description{
Provides a concise summary of \code{nicheR} objects.
}
\details{
The function formats and rounds key quantities for readability but does
not modify the underlying object.
}
\seealso{
\code{\link{build_ellipsoid}}
}
