% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_comunities.R
\name{random_ellipses}
\alias{random_ellipses}
\title{Generate random ellipses constrained by a point cloud and a reference ellipse}
\usage{
random_ellipses(
  object,
  background,
  n = 10,
  smaller_proportion = 0.1,
  uniform_centroids = TRUE,
  resolution = 50,
  seed = 1
)
}
\arguments{
\item{object}{A nicheR_ellipsoid object used as a reference ellipse
(the biggest to be generated), and containing at least
\code{covariance_matrix} and \code{cl}.}

\item{background}{Matrix or Dataframe. The 2D point cloud (coordinates)
used to select random centroids.}

\item{n}{Integer. Number of ellipses to generate.}

\item{smaller_proportion}{Numeric. Minimum scaling factor for the covariance.
Must be between 0 and 1. Default = 0.1. This controls how much smaller
the new ellipses can be compared to the reference.}

\item{uniform_centroids}{Logical. If TRUE, centroids are sampled more
uniformly across the background using a grid-based thinning approach.
Default = TRUE.}

\item{resolution}{Integer. Number of cells per side in the grid to deal with
point density variation across background.}

\item{seed}{Integer. Random seed for reproducibility. Default = 1.
Set to NULL for no seeding.}
}
\value{
A list of length \code{n} with ellipse features.
}
\description{
Creates n random ellipses with centroids sampled from an irregular point
cloud. Covariance matrices are built using random rotations and scaled
eigenvalues restricted by user-defined limits.
}
