% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_read_nicheR.R
\name{read_nicheR}
\alias{read_nicheR}
\title{Read a nicheR object from disk}
\usage{
read_nicheR(file)
}
\arguments{
\item{file}{Character. The path to the file to be read.}
}
\value{
The saved \code{nicheR_ellipsoid} or \code{nicheR_community} object.
}
\description{
A wrapper around \code{\link[base]{readRDS}} to load saved nicheR
objects back into the R environment.
}
\examples{
# Build and save an ellipsoid first
range_df <- data.frame(bio_1  = c(15, 25),
                       bio_12 = c(500, 1500))
ell <- build_ellipsoid(range = range_df)

tmp <- tempfile(fileext = ".rds")
save_nicheR(ell, file = tmp)

# Read it back
ell_loaded <- read_nicheR(tmp)
ell_loaded

\dontshow{file.remove(tmp)}

}
\seealso{
\code{\link{save_nicheR}}
}
