% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_read_nicheR.R
\name{read_nicheR}
\alias{read_nicheR}
\title{Read a nicheR object from disk}
\usage{
read_nicheR(file)
}
\arguments{
\item{file}{Character. The path to the file to be read.}
}
\value{
The saved \code{nicheR_ellipsoid} or \code{nicheR_community} object.
}
\description{
A wrapper around \code{\link[base]{readRDS}} to load saved nicheR
objects back into the R environment.
}
\seealso{
\code{\link{save_nicheR}}
}
