% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{sample_data}
\alias{sample_data}
\title{Sample Occurrence Data from a Prediction Surface}
\usage{
sample_data(
  n_occ,
  prediction,
  prediction_layer = NULL,
  sampling = "centroid",
  method = "suitability",
  sampling_mask = NULL,
  seed = 1,
  verbose = TRUE
)
}
\description{
Sample Occurrence Data from a Prediction Surface
}
