% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_read_nicheR.R
\name{save_nicheR}
\alias{save_nicheR}
\title{Save a nicheR object to disk}
\usage{
save_nicheR(object, file, overwrite = FALSE, ...)
}
\arguments{
\item{object}{A \code{nicheR_ellipsoid} or \code{nicheR_community} object.}

\item{file}{Character. The connection or name of the file where the
object will be saved (usually ending in ".rds").}

\item{overwrite}{Logical. If \code{TRUE}, an existing file at the
specified path will be replaced. Default is \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link[base]{saveRDS}}.}
}
\value{
No return value. Saves the object to the specified path.
}
\description{
A wrapper around \code{\link[base]{saveRDS}} to save \code{nicheR_ellipsoid}
or \code{nicheR_community} objects to a file. Includes a safety check
for overwriting existing files and ensures the file extension is .rds.
}
