% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_theta.R
\name{start_theta}
\alias{start_theta}
\title{Starting values for niche model on math scale}
\usage{
start_theta(env_occ)
}
\arguments{
\item{env_occ}{Data frame with environmental values at presence points.}
}
\value{
Numeric vector of starting values for `theta`.
}
\description{
Starting values for niche model on math scale
}
\examples{
start_theta(example_env_occ_2d)
}
