% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_theta_multiple.R
\name{start_theta_multiple}
\alias{start_theta_multiple}
\title{Generate multiple starting points for niche model optimization}
\usage{
start_theta_multiple(
  env_data,
  num_starts = 100,
  quant_vec = c(0.1, 0.5, 0.9),
  method = "sobol"
)
}
\arguments{
\item{env_data}{Data frame of environmental values from the accessibility
area M.}

\item{num_starts}{Integer. Number of starting points.}

\item{quant_vec}{Quantiles for mu ranges.}

\item{method}{Character: "sobol" or "uniform".}
}
\value{
A data frame with num_starts rows and columns corresponding to parameters.
}
\description{
Generate multiple starting points for niche model optimization
}
