% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_data.R
\name{virtual_data}
\alias{virtual_data}
\title{Generate data within based on a ellipsoidal niche}
\usage{
virtual_data(object, n = 100)
}
\arguments{
\item{object}{A \code{nicheR_ellipsoid} object containing at least
\code{centroid} and \code{cov_matrix}.}

\item{n}{Integer. The number of virtual points to generate. Default = 100.}
}
\value{
A matrix with \code{n} rows and columns corresponding to the
dimensions of the input ellipsoid.
}
\description{
Simulates \code{n} random points from a multivariate normal distribution
defined by the centroid and covariance matrix of a \code{nicheR_ellipsoid}
object.
}
\details{
The function uses eigen-decomposition to transform standard normal
variables into the coordinate system defined by the ellipsoid's
covariance structure.
}
